@JayCesar | feat: implements account methods from interface and creates AccountProducer and AccountRequester files

+++ MessageFlow/src/main/java/com/ibmmq/messageflow/repository/AccountRepository.java
@@ -15,9 +15,11 @@ public interface AccountRepository {
     public void delete(int number);
 
     // Banking Methods
-    public void withdraw(int number, float amount);
+    public void withdraw(int number, double amount);
 
-    public void deposit(int number, float amount);
+    public void deposit(int number, double amount);
 
-    public void transfer(int sourceNumber, int destinationNumber, float amount);
+    public void transfer(int sourceNumber, int destinationNumber, double amount);
+
+
 }
